==> 2-feature-extraction-all.ipynb <==
### 1. Import the essential libraries

import os
import sys
import cv2
import random
import joblib
import numpy as np
import pandas as pd
from matplotlib import pyplot as plt

from sklearn.cluster import KMeans
from sklearn import metrics
from scipy.spatial.distance import cdist

from skimage.color import rgb2lab, lab2rgb
from skimage.segmentation import slic
from skimage.segmentation import mark_boundaries
from skimage.util import img_as_float

from tqdm import tqdm_notebook

### 1.1 Define some helpful functions

def surf_feature_extraction_v2(image, indices, hessianThreshold=400, nOctaves=3, n_keypoints=10, surf_window=20):
    # construct a mask for the segment
    surf = cv2.xfeatures2d.SURF_create(hessianThreshold=hessianThreshold, nOctaves=nOctaves)
    key_points = [cv2.KeyPoint(y, x, surf_window) for (x,y) in indices]
    key_points = sorted(key_points, key=lambda x: -x.response)[:n_keypoints]
    key_points, description = surf.compute(image, key_points)
    
    if len(key_points) != 0:
        description = description.flatten()
    else:
        description = np.array([])
        
    fearute_vector_size = (n_keypoints * 64)
    if description.size < fearute_vector_size:
        description = np.concatenate([description, np.zeros(fearute_vector_size - description.size)])
    
    return (key_points, description)

def calc_energy_and_amplitude(filtered_img):
    '''
    Return a list of values:
      * Local Energy = summing up the squared value of each matrix value from a response matrix
      * Mean Amplitude = sum of absolute values of each matrix value from a response matrix
    '''
    local_energy = np.sum(np.square(filtered_img))
    mean_amplitude = np.sum(np.absolute(filtered_img))
    return [local_energy, mean_amplitude]

def gabor_feature_extraction_v2(image_lab_L, theta_range=[0, np.pi/6, np.pi/4, np.pi/3, np.pi/2, 2*np.pi/3, 3*np.pi/4, 5*np.pi/6], scale_range=[3, 6, 13, 28, 58]):
    sp_local_energy = []
    sp_mean_amplitude = []
    for scale in scale_range:
        for angle in theta_range:
            g_kernel = cv2.getGaborKernel(ksize=(20, 20), sigma=scale, theta=angle, lambd=7, gamma=0.9, psi=1.5, ktype=cv2.CV_32F)
            filtered_img = cv2.filter2D(image_lab_L, cv2.CV_8UC3, g_kernel)
            
            lcl_enrgy, mn_amplitude = calc_energy_and_amplitude(filtered_img)
            
            sp_local_energy.extend([lcl_enrgy])
            sp_mean_amplitude.extend([mn_amplitude])
            
    return (sp_local_energy, sp_mean_amplitude)

### 2. Set the paths for *train* and *test* data

train_images_path = os.path.join('.', 'data', 'Train')
res_path          = os.path.join('.', 'res_all')

### 2.1. Get the Train Images & the Color Palette

train_images = joblib.load(os.path.join(res_path,'trainImages.joblib'))
colorPalette = joblib.load(os.path.join(res_path,'colorPalette.joblib'))
print (f'{type(train_images)}\n{type(colorPalette)}')

np.sort(train_images)

### 3. Get Images and Perform the Preprocessing (cont.)
  * #### 3.1: Read Image
  * #### 3.2: Normalize Values from [0, 255] to [0, 1]
  * #### 3.3: Convert Image from RGB to LAB
  * #### 3.4: Get the Superpixels for an Image using the SLIC Segmentation Algorithm
  * #### 3.5: Extract the SURF Features from an Image per Superpixel
  * #### 3.6: Extract the Gabor Features from an Image per Superpixel
  * #### 3.7: Get the Dominant Color per Superpixel
  * #### 3.8: Associate each feature vector {SURF, Gabor} with the corresponding color class of the dominant color in the color palette

dataset = None

for img in tqdm_notebook(train_images):
    # 3.1: Read Image
    trainImage = cv2.imread(os.path.join(train_images_path, img))
    trainImage = cv2.cvtColor(trainImage, cv2.COLOR_BGR2RGB)
    # 3.2: Normalize Values from [0, 255] to [0, 1]
    trainImageScaled = trainImage/255
    # 3.3: Convert Image from RGB to LAB
    trainImageLAB = rgb2lab(trainImageScaled)
    # 3.4: Get the Superpixels for an Image using the SLIC Segmentation Algorithm
    segments = slic(trainImageLAB, convert2lab=False, n_segments=600, sigma=5)
    # 3.5: Extract the SURF Features from an Image per Superpixel
    # 3.6: Extract the Gabor Features from an Image per Superpixel
    # 3.7: Get the Dominant Color per Superpixel
    for (i, segment) in enumerate(np.unique(segments)):
#         print (f'Processed superpixel: {segment}\t Color Class: {color_vector}\r', end='')Z
        # Construct a mask for the segment
        mask = np.zeros(trainImage.shape[:2], dtype = "uint8")
        mask[segments == segment] = 255
        superpixel = cv2.bitwise_and(trainImageLAB, trainImageLAB, mask=mask)
                
        superpixel_colors = pd.DataFrame(np.vstack(superpixel[segments==segment, 1:3]), columns=['a', 'b'])    
#         superpixel_colors = pd.DataFrame(np.vstack(superpixel[:, :, 1:3]), columns=['a', 'b'])    

        maximal_color = KMeans(n_clusters=5, n_init=4, max_iter=100, n_jobs=-1)
        maximal_color.fit(superpixel_colors.values)
        superpixel_colors['y_color'] = maximal_color.labels_
        superpixel_dominant_color = superpixel_colors.loc[superpixel_colors['y_color'] == superpixel_colors['y_color']  
                                                     .value_counts().index[0]].apply(lambda x: x.median())[['a', 'b']]  
#         superpixel_dominant_color = superpixel_colors.loc[superpixel_colors['y_color'] == superpixel_colors['y_color']  
#                                                      .value_counts().index[1]].apply(lambda x: x.median())[['a', 'b']]  

        superpixel_dominant_color_class = colorPalette.predict(superpixel_dominant_color.values.reshape(1,-1))
          
        surf_vector = surf_feature_extraction_v2(np.uint8(trainImageLAB[:,:,0]), np.argwhere(segments == segment), n_keypoints=7, surf_window=20)[1] 
        gabor_vector = np.hstack(gabor_feature_extraction_v2(np.uint8(superpixel[:,:,0])))
        color_vector = np.array(superpixel_dominant_color_class[0])
        feature_vector = np.hstack((surf_vector, gabor_vector, color_vector))
        
        if (dataset is None):
            dataset = np.array(feature_vector)
        else:
            dataset = np.row_stack((dataset, feature_vector))

pdDataset = pd.DataFrame(dataset)
pdDataset

# CHECKPOINT
* ### Save the train image dataset for the following notebooks

if not os.path.exists(os.path.join('.', 'res_all')):
    os.makedirs(os.path.join('.', 'res_all'))
    pdDataset.to_pickle(os.path.join(save_path, 'pdDataset.pkl'))
save_path = os.path.join('.', 'res_all')

pdDataset.to_pickle(os.path.join(save_path, 'pdDataset.pkl'))
