==> 3-svm-training.ipynb <==
### 1. Import the essential libraries

import os
import sys
import cv2
import random
import joblib
import numpy as np
import pandas as pd
from matplotlib import pyplot as plt

from sklearn.cluster import KMeans
from sklearn import metrics
from scipy.spatial.distance import cdist

from skimage.color import rgb2lab, lab2rgb
from skimage.segmentation import slic
from skimage.segmentation import mark_boundaries
from skimage.util import img_as_float

params = {"ytick.color" : "yellowgreen",
          "xtick.color" : "yellowgreen",
          "axes.labelcolor" : "yellowgreen",
          "axes.edgecolor" : "yellowgreen"}
plt.rcParams.update(params)

### 2. Set the paths for *train* and *test* data

train_images_path = os.path.join('.', 'data', 'Train')
res_path          = os.path.join('.', 'res_all')

### 2.1. Get the Train Image DataFrame & the Color Palette

colorPalette = joblib.load(os.path.join(res_path, 'colorPalette.joblib'))
type(colorPalette)

imageDataFrame = pd.read_pickle(os.path.join(res_path, 'pdDataset.pkl'))
imageDataFrame

### 3. Dataset Preprocessing
* #### 3.1. Importing the Essential Libraries
* #### 3.2. Standard Scaling (mean=0, std=1)
* #### 3.3. Dimensionality Reduction 

from sklearn.model_selection import RepeatedKFold, RepeatedStratifiedKFold
from sklearn.preprocessing import StandardScaler, MinMaxScaler
from sklearn.metrics import accuracy_score
from tqdm import tqdm_notebook
from sklearn.model_selection import train_test_split
from sklearn.multiclass import OneVsRestClassifier
from sklearn.decomposition import PCA
from sklearn.svm import SVC, LinearSVC

pca = PCA(n_components=64, svd_solver='auto', random_state=100)
stdScaler = StandardScaler()

X_scaled = stdScaler.fit_transform(imageDataFrame.iloc[:,:-1].values)
X_pca = pca.fit_transform(X_scaled)

print(f'''\t\t\tPCA Dimensionality Reduction
        \nExplained Variance Ratio:\n{pca.explained_variance_ratio_}
        \nSum: {np.sum(pca.explained_variance_ratio_)}''')  

dataset = pd.DataFrame(np.column_stack((X_pca, imageDataFrame.iloc[:,-1].values)))

dataset.head(10)

### 4. SVM Training
  * #### One vs All Method (i.e: We need to create as many SVM's as the Colors in our Palette)
  * #### With the help of Repeated k-Fold Validation, for *n_repeats* rounds we:
    * ##### Fragment our dataset to *n_splits* random fragments
    * ##### Train our set of SVMs

clf = SVC(kernel='rbf', gamma='scale', cache_size=700, random_state=110)
rkf = RepeatedKFold(n_splits=10, n_repeats=2, random_state=100)
ovr_clf = OneVsRestClassifier(clf, n_jobs=-1)

X = dataset.iloc[:,:-1].values
y = dataset.iloc[:,-1].values
accuracy = np.array([])

for train_index, test_index in tqdm_notebook(rkf.split(X), total=20):
    X_train, X_test = X[train_index], X[test_index]
    y_train, y_test = y[train_index], y[test_index]
    ovr_clf.fit(X_train, y_train)
    acc = ovr_clf.score(X_test,y_test)
    print (f'Accuracy: {acc}')
    accuracy = np.append(accuracy, [acc])

print(f'Max. Accuracy: {np.amax(accuracy)}')
print(f'Min. Accuracy: {np.amin(accuracy)}')

ovr_clf.classes_

# CHECKPOINT
* ### Save the OvR SVM Classifier for the following notebooks

if not os.path.exists(os.path.join('.', 'res_all')):
    os.makedirs(os.path.join('.', 'res_all'))
    
save_path = os.path.join('.', 'res_all')

joblib.dump(ovr_clf, os.path.join(save_path,'TextureColorClassifierOvR.joblib'))
# joblib.dump(clf, os.path.join(save_path,'TextureColorClassifierOvR.joblib'))

joblib.dump(pca, os.path.join(save_path,'DecompositionPCA.joblib'))

joblib.dump(stdScaler, os.path.join(save_path,'stdScaler.joblib'))
